Reorganize files into a python module
