+++ Dockerfile
@@ -4,5 +4,5 @@ COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
-
-ENTRYPOINT [ "python", "./chrome.py" ]
+COPY chrome2mqtt/ chrome2mqtt/
+ENTRYPOINT [ "python", "-m", "chrome2mqtt" ]
